Total time changed to decimal in gamesessions migration.

+++ database/migrations/2015_12_01_193928_create_gamesessions_table.php
@@ -28,7 +28,7 @@ class CreateGamesessionsTable extends Migration
             $table->integer('attempt_num');
             $table->timestamp('start_time');
             $table->timestamp('end_time');
-            $table->integer('total_time');
+            $table->double('total_time');
             $table->integer('moves');
         });
     }
